refactor: separate the runtime's stateless helpers from its orchestration
deepagents_runtime.py held the class that sequences a turn and every helper
that class calls in one 6,611-line file. This moves the helpers out: catalog
scope reasoning, evidence shaping, response assembly, diagnostics, and the
pydantic schemas for tool arguments. None of them touches self, the graph, or a
live service, so they can now be read and tested without standing up a runtime.

The helpers move as one module rather than as themed ones because the call
graph between those themes is cyclic -- evidence calls scope, scope calls
shared helpers, and those call back into diagnostics and response assembly.
Splitting further means breaking those cycles, which is a design change rather
than a move, so it is left to a separate change instead of smuggled into this one.

No behaviour changes: every one of the 141 moved definitions is byte-identical
to its merge-base text, and the 16 that stay behind are byte-identical too.
