+++ tests/unit/chain_server/test_deepagents_observability.py
@@ -20,6 +20,8 @@ from .tool_evidence_fixtures import (
 )
 from chain_server.src.deepagents_runtime import (
     DeepAgentsRuntime,
+)
+from chain_server.src.turn_support import (
     RequestIdentity,
     _collect_agent_diagnostics,
 )
